refactor:: data class Hero.kt
