+++ app/src/main/java/com/example/marvel/data/Hero.kt
@@ -1,0 +1,5 @@
+package com.example.marvel.data
+
+import com.example.marvel.MarvelScreen
+
+data class Hero(val logo: String, val name: String, val description: String, val screen: MarvelScreen)
